requirements.txt
python-dotenv
google-generativeai
Pillow
ollama
pytesseract
easyocr
python-doctr[torch]